deletes old manifest file under the testnet package
